requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
frozenlist==1.6.0
gTTS==2.5.4
idna==3.10
iso8601==2.1.0
multidict==6.4.3
propcache==0.3.1
pydub==0.25.1
pygame==2.6.1
python-dotenv==1.1.0
requests==2.32.3
twitchio==2.10.0
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
